// src/app/core/api/generated/model/animalCreateDTO.ts
/**
 * VetClinicAPI_V2 | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GetAnimalByIdIdParameter } from './getAnimalByIdIdParameter';


export interface AnimalCreateDTO { 
    code: string;
    name: string;
    species: number;
    dateOfBirth: string;
    gender: number;
    ownerId: GetAnimalByIdIdParameter;
}