// src/app/core/api/generated/model/ownerResponseDTO.ts
/**
 * VetClinicAPI_V2 | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GetAnimalByIdIdParameter } from './getAnimalByIdIdParameter';
import { AnimalSummaryDTO } from './animalSummaryDTO';


export interface OwnerResponseDTO { 
    id?: GetAnimalByIdIdParameter;
    name?: string;
    phone?: string;
    animals?: Array<AnimalSummaryDTO>;
}